// app/models/resturants.ts
export class Resturant {
    resturantId : number = 0;
    resturantName!: string ;
    resturantTypeDish: string = "";
    resturantRating: number = 0;
    resturantDistance: number = 0;
    resturantImg: string = "";
    resturantDishPrice: number = 0;
    resturantDistanceTime: number =0;
    deliveryfee: number = 50;
    orderquantity:number = 0;
}